+++ app/jni/src/App/AppGfx.h
@@ -14,7 +14,7 @@
 #include "GameSettings.h"
 #include "Languages.h"
 #include "Traits.h"
-#include "WinTypeGlobal.h"
+#include "TypeGlobal.h"
 
 class MusicManager;
 class HighScore;
