+++ app/jni/src/App/GameSettings.cpp
@@ -15,7 +15,7 @@
 
 #include "Config.h"
 #include "MusicManager.h"
-#include "WinTypeGlobal.h"
+#include "TypeGlobal.h"
 
 #if PLATFORM_EMS
 #include <emscripten.h>
