+++ app/jni/src/App/HighScore.cpp
@@ -8,7 +8,7 @@
 #include "GameSettings.h"
 #include "GfxUtil.h"
 #include "MusicManager.h"
-#include "WinTypeGlobal.h"
+#include "TypeGlobal.h"
 
 #if PLATFORM_EMS
 #include <emscripten.h>
