+++ app/jni/src/App/MusicManager.h
@@ -4,7 +4,7 @@
 #include <SDL3_mixer/SDL_mixer.h>
 
 #include "ErrorInfo.h"
-#include "WinTypeGlobal.h"
+#include "TypeGlobal.h"
 
 class MusicManager {
    public:
