+++ app/jni/src/App/TypeGlobal.h
