+++ app/jni/src/CompGfx/KeyboardGfx.h
@@ -8,7 +8,7 @@
 
 #include "ButtonGfx.h"
 #include "Traits.h"
-#include "WinTypeGlobal.h"
+#include "TypeGlobal.h"
 
 using namespace traits;
