+++ app/jni/src/CompGfx/LabelLinkGfx.h
@@ -9,7 +9,7 @@
 #include <SDL3_ttf/SDL_ttf.h>
 
 #include "Traits.h"
-#include "WinTypeGlobal.h"
+#include "TypeGlobal.h"
 
 using namespace traits;
